change class malloc to new

+++ LSTable.h
@@ -27,7 +27,7 @@ public:
   void update_by_ls(char* packet, unsigned int current_time, unsigned short size);
   void dijkstra(hash_map<unsigned short, unsigned short>& routing_table);
   void set_ls_packet(char* packet, unsigned short packet_size);
-  bool update_by_pong(unsigned short src_id, unsigned short cost, unsigned short current_time);
+  bool update_by_pong(unsigned short src_id, unsigned short cost, unsigned int current_time);
   void increase_seq() { ++sequence_num; }
   void delete_neighbor(unsigned short neighbor_id);
   void print_routing_table(hash_map<unsigned short, unsigned short>& routing_table);
